relocate getStakerConfigByNetwork
