+++ packages/dappmanager/src/api/routes/metrics.ts
@@ -1,7 +1,7 @@
 import client from "prom-client";
 import { wrapHandler } from "../utils.js";
 import * as db from "../../db/index.js";
-import { getStakerConfigByNetwork } from "../../modules/stakerConfig/getStakerConfigByNetwork.js";
+import { getStakerConfigByNetwork } from "../../modules/stakerConfig/index.js";
 import { listPackageNoThrow } from "../../modules/docker/list/index.js";
 import { isEmpty } from "lodash-es";
 import { Network } from "@dappnode/types";
