+++ packages/dappmanager/src/modules/stakerConfig/get/getStakerConfig.ts
@@ -19,7 +19,7 @@ import { ReleaseFetcher } from "../../release/index.js";
 import { getBeaconServiceName, pickStakerItemData } from "../utils.js";
 import { Network } from "@dappnode/types";
 import { getStakerDnpNamesByNetwork } from "./getStakerDnpNamesByNetwork.js";
-import { getStakerConfigByNetwork } from "../getStakerConfigByNetwork.js";
+import { getStakerConfigByNetwork } from "../index.js";
 import { eventBus } from "../../../eventBus.js";
 import * as db from "../../../db/index.js";
