+++ packages/dappmanager/src/modules/stakerConfig/index.ts
@@ -1,3 +1,3 @@
 export * from "./get/index.js";
 export * from "./set/index.js";
-export { getStakerConfigByNetwork } from "./getStakerConfigByNetwork.js";
+export { getStakerConfigByNetwork } from "./utils.js";
