+++ packages/dappmanager/src/modules/stakerConfig/utils.ts
@@ -1,14 +1,44 @@
 import {
   UserSettingsAllDnps,
-  InstalledPackageData,
-  InstalledPackageDataApiReturn,
   StakerItemData,
-  PackageRelease
+  PackageRelease,
+  ConsensusClient,
+  ExecutionClient,
+  StakerConfigByNetwork
 } from "@dappnode/common";
-import { logs } from "../../logs.js";
-import { dockerContainerStop } from "../docker/index.js";
 import { pick } from "lodash-es";
 import { Manifest, Network } from "@dappnode/types";
+import * as db from "../../db/index.js";
+
+export function getStakerConfigByNetwork<T extends Network>(
+  network: T
+): StakerConfigByNetwork<T> {
+  switch (network) {
+    case "mainnet":
+      return {
+        executionClient: "geth.dnp.dappnode.eth" as ExecutionClient<T>,
+        consensusClient: db.consensusClientMainnet.get() as ConsensusClient<T>,
+        feeRecipient: db.feeRecipientMainnet.get(),
+        isMevBoostSelected: db.mevBoostMainnet.get()
+      };
+    case "gnosis":
+      return {
+        executionClient: db.executionClientGnosis.get() as ExecutionClient<T>,
+        consensusClient: db.consensusClientGnosis.get() as ConsensusClient<T>,
+        feeRecipient: db.feeRecipientGnosis.get(),
+        isMevBoostSelected: false // gnosis doesn't support mevBoost
+      };
+    case "prater":
+      return {
+        executionClient: db.executionClientMainnet.get() as ExecutionClient<T>,
+        consensusClient: db.consensusClientMainnet.get() as ConsensusClient<T>,
+        feeRecipient: db.feeRecipientMainnet.get(),
+        isMevBoostSelected: db.mevBoostPrater.get()
+      };
+    default:
+      throw new Error(`Network ${network} not supported`);
+  }
+}
 
 /**
  * Get the validator service name.
